Move tab components to navigation folder
